Resolving build
